fixed login signup broken functionality
