+++ assets/html/login.html
@@ -21,7 +21,7 @@
   <script type="module" src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js"></script>
   <script nomodule src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.js"></script>
 
-  <script type="module" src="../js/loginSignUp.js" defer></script>
+  <script type="module" src="../js/login.js" defer></script>
   <script src="../js/rateUsModal.js" defer></script>
 
   <!-- rroushan verma -->
@@ -339,7 +339,7 @@
     let eyeIconLogin = document.getElementById("eyeicon-login");
     let passwordLogin = document.getElementById("login_password");
     let eyeIcon = document.getElementById("eyeicon");
-    let password = document.getElementById("password");
+    let password = document.querySelector("#password");
     let passwordStrength = document.getElementById("password-strength");
     let lowUpperCase = document.querySelector(".low-upper-case i");
     let number = document.querySelector(".one-number i");
