+++ assets/html/mylogin.html
@@ -16,6 +16,7 @@
     <script type="module" src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.esm.js"></script>
     <script nomodule src="https://unpkg.com/ionicons@5.5.2/dist/ionicons/ionicons.js"></script>
     <script src="../js/signinGoogle.js" defer type="module"></script>
+    <script src="../js/signup.js" defer type="module" ></script>
 </head>
 
 <body>
@@ -125,13 +126,12 @@
             <input type="checkbox" id="chk" aria-hidden="true" />
 
             <div class="signup">
-                <form class="form" style="padding: 15px 35px;">
+                <form class="form signup-form" style="padding: 15px 35px;">
                     <label for="chk" aria-hidden="true">Sign up</label>
                     <input class="input" type="text" name="txt" placeholder="User name" required="" />
-                    <input class="input" type="email" name="email" placeholder="Email" required="" />
-                    <input class="input" type="password" name="pswd" placeholder="Password"  required="" />
-                    
-                    <button>Sign up</button>
+                    <input class="input" type="email" name="email" placeholder="Email" id="email" required="" />
+                    <input class="input" type="password" name="pswd" placeholder="Password" id="password"  required="" />
+                    <button type="submit">Sign up</button>
                     <div id="or">Or</div>
                     <button id="google-sign-in"> Sign Up with Google </button>
                 </form>
